backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
Pillow==10.3.0
python-dotenv==1.0.1
groq==0.9.0
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
requests==2.32.3
ffmpeg-python==0.2.0
tiktoken==0.7.0
reportlab==4.2.0
